requirements.txt
Flask==2.0.1
Pillow==8.4.0
pandas==1.3.3
gunicorn==20.1.0
numpy==1.24.4
Flask-RESTful==0.3.9
SQLAlchemy==1.4.32
psycopg2==2.9.2
pytest==7.4.4
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
matplotlib==3.7.4